refactor: move student history logic to separate module
